Add support for defining schemas to be used by the sinks
Right now the Kafka sink does not support the use of the defined
schemas, but these allow for defining valid/acceptable schemas up front
for data written to specific topics.

What this will _not_ do however is any form of type coercion! Make sure
the schemas are the right types for the data coming in!